+++ packages/core/src/index.ts
@@ -25,3 +25,4 @@ export * from "./worker.js";
 export * from "./guardians.js";
 export * from "./conversations.js";
 export * from "./spawn.js";
+export * from "./usage.js";